Move from gcr.io/cri-tools to gcr.io/k8s-staging-cri-tools
This moves the images from `gcr.io/cri-tools` to the community
maintained project `gcr.io/k8s-staging-cri-tools` as fix for many broken
CI systems.

I'm not able to build and push the windows images for now, which leaves
them untouched in this patch.

As a follow-up, we have to put the same automation in place for
`images/hostnet-nginx` and `images/image-test` as we have it for
`images/image-user`.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>